Have the #include directives compatible with buried headers
When the SOCI library is used as an API by other programs, the
headers are better be buried, i.e., prefixed with specific
paths.
That way, there is no ambiguity when including SOCI headers.
Moreover, there is no need to add specific compilation flags
such as '-I/usr/include/soci'.